move linuxplugin interface-related code to linuxplugin/ifplugin
