+++ plugins/linuxplugin/ifplugin/ifaceidx/doc.go
