+++ plugins/linuxplugin/ifplugin/interface_utils.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package linuxplugin
+package ifplugin
 
 import (
 	"net"
